Updated webpage to add grid button and fix drawing behavior for grids
Previous drawing behavior for grids was incorrect. Pixels were being
positioned absolutely to body because parent grids did not have the
"Position: relative" property. A quirk of JS, not too sure how it works.
Furthermore, updated behavior to place the pixels based on the detected
offset of the mouse pointer's location when the mouseEvent is triggered.
Because the pixels were being positioned absolutely, this messed with
the read offset numbers: fixed by setting pixel pointer-events to none.