Change approach of grid building
I accidentally misinterpreted the assignment and attempted to do an
approach with a single container and flex wrapping. This commit pulls
from the "calc-approach" and salvages some of the logic to move
forward with this new grid building appraoch. This approach first builds
a row with a loop and then builds each column with another loop.
TODO: figure out how to make sure grid stay within box.